Feat: 분석 뉴스 리스트 실행 구현
Former-commit-id: accc69505409ebe82f90715495f0906a3e3aea85

+++ android/app/src/main/java/com/ssafy/firstproject/ui/home/adpater/NewsAdapter.kt
@@ -41,7 +41,7 @@ class NewsAdapter(private val itemClickListener: ItemClickListener) :
             binding.tvTitle.text = item.title
             binding.tvSummary.text = item.description
             binding.tvTruthPercent.text =
-                binding.root.context.getString(R.string.reliability, item.reliability)
+                binding.root.context.getString(R.string.reliability, item.reliability.toInt())
 
             binding.root.setOnClickListener {
                 itemClickListener.onClick(item.id)
